Fixed generator to get along with django
Generator now uses the Room class from models.py, which brings in the payer relevant functions. Room class in models.py now has a more succinct room connector function.
