+++ adventure/flavor_lists.py
@@ -1,8 +1,8 @@
-houses_list = ["Papa", "Jokey", "Brainy", "Clumsy", "Grouchy", "Hefty", "Greedy",
+houses = ["Papa", "Jokey", "Brainy", "Clumsy", "Grouchy", "Hefty", "Greedy",
         "Handy", "Baby", "Farmer", "Baker", "Vain", "Finance", "Painter",
         "Poet", "Harmony", "Gutsy", "Chef", "Narrator", "Nosey", "Panicky",
         "Crazy"]
-buildings_list = ["Church", "General Store", "Garden", "Town Hall", "Library",
+buildings = ["Church", "General Store", "Garden", "Town Hall", "Library",
     "Secret hole", "Prison", "School", "Office", "Gym", "Post office", "Fire department",
     "Restaurant", "Unholy Alter to the Profane God Grgammth", "Park", "Playground"]
 desc_start = ["It looks", "This buiding is remarkably"]
@@ -10,7 +10,7 @@ desc_adj = ["tall", "rich looking", "ugly", "vandalized"]
 desc_mid = ["There is a", "It could use a"]
 desc_noun = ["light", "groundskeeper", "window", "new coat of paint"]
 desc_end = ["out back", "", "near the front door"]
-extra_adjectives_list = ["amused",
+extra_adjectives = ["amused",
 "chivalrous",
 "fearless",
 "common",
